cleaned up percent
